refactor: split coloringImages.tsx (354 lines) into per-image files (#83)
Each of the 8 SVG coloring images (cat, house, sun, butterfly, flower,
fish, tree, car) now lives in its own file under components/images/.
imageComponents.ts imports directly from each file. Adding a new
coloring image now means adding one ~40-line file.
